Merge work from previous few months

+++ include/simulation/ray_processor.hpp
@@ -48,7 +48,7 @@ public:
   virtual void process_ray(unsigned scope_id, const calin::math::ray::Ray& ray,
     double pe_weight);
   virtual void finish_processing();
-  unsigned process_all_from_ray_generator(
+  uint64_t process_all_from_ray_generator(
     calin::math::ray_generator::RayGenerator* gen, unsigned scope_id = 0);
 };
